+++ src/ArrayBuilder.ts
@@ -15,9 +15,9 @@ export default class ArrayBuilder<T> extends CollectionBuilder<ArrayOf<T>> {
         const nextValue = [...value];
 
         value.forEach((item: T, i: number) => {
-          nextValue[i] = contents.runChecks(
-            `${path}[${i}]`,
+          nextValue[i] = contents.run(
             item,
+            `${path}[${i}]`,
             this.currentStruct,
             this.options,
           )!;